+++ CHANGELOG.md
@@ -14,5 +14,12 @@
 - Icon for extension
 - Tested on MacOS & Windows
 
-## [1.0.0] - Release
+## 1.0.0 - Release
 - Initial release
+
+## 1.1.0 - Clean up command
+- Users can now "clean up" their SDF XML objects
+    - Users will select which files they want to "clean"
+    - Users will then be prompted to create a new sub-folder under `src/Objects` or place them into an existing folder
+    - Files will be automatically moved into the chosen sub-folder, and objects will be placed into meaningful folders
+        - e.g: An object which represents a custom field for Sales Orders will be automatically placed into `Transaction Body Custom Fields`